+++ src/modules/powerdisplay/PowerDisplayModuleInterface/PowerDisplayProcessManager.h
@@ -50,7 +50,8 @@ private:
     HRESULT start_process(const std::wstring& pipe_name);
     HRESULT start_named_pipe_server(const std::wstring& pipe_name);
     void refresh();
-    void send_named_pipe_message(const std::wstring& message_type, const std::wstring& message_arg = L"");
+    bool restart_pipe();
+    HRESULT send_named_pipe_message(const std::wstring& message_type, const std::wstring& message_arg = L"");
 
     OnThreadExecutor m_thread_executor; // all internal operations are done on background thread with task queue
     std::atomic<bool> m_enabled = false; // written on main thread, read on background thread